Move GlobalMemoryBuffer into allocators.py
GlobalMemoryBuffer and get_global_memory_buffer are allocation primitives
(RotaryBucketAllocator routes every tensor request through them), and
nothing outside the megatron_fsdp subpackage imports them from .utils.
Relocate both into allocators.py next to the bucket allocators, drop the
now-unused operator/reduce/nullcontext imports from utils.py, and update
param_and_grad_buffer.py to import get_global_memory_buffer from
.allocators.

No behavior change — pure code move.
